Check FLAG_CLEAR_BUF behavior.
This checks that the kernel actually clears the buffer space for a reply
parcel as requested. No check is made that the deallocated memory
is also cleared since it is deallocated, and we have no control over
what happens to it. Ramdumps should be used to verify use of this flag,
but this test is a basic sanity check for them. In the future, we could
consider expanding this by adding something like libhwbinder's
postCommandTask here. However, for now, avoiding overhead for test only.

Bug: 171501998
Test: binderClearBufTest
Change-Id: I797a9e9b6a9b8c70816ce51bb9bc22f518b54f77
Signed-off-by: Akash Srivastava <akashniki@gmail.com>